rclone-backup: Add "--newest" command to print newest source file
Add new script mode `--newest`, which joins `--ls` and `--verify` as
alternatives to doing a backup. In this mode the script scans the
source directory to find the newest file and prints the timestamp and
the path of that file. This can be used by backup scripts to help
determine which files should be verified after a backup.