mongo-incremental-export.py: Don't re-export documents unnecessarily!
We were relying on Python always ordering the dicts returned by
MongoDB the same, which was a bad assumption. As a result, sometimes
documents that were not modified were being exported again because
their fields ended up being serialized in a different order so their
checksums were different.

We now use `SON` objects for representing the data sent to/from the
server, so order is preserved, so we won't export documents that don't
need to be exported.